- Fixed an error
In the addTask() function, the variable taskTitle = capitalizeTitle(UI_Elements["task-title"].value); can result in undefined if the user leaves the input blank. This happens because capitalizeTitle() is called with an empty or undefined value, leading to an error.

To fix this, I change the logic that if the user leaves the input blank ("") the variable taskTitle would store an empty string, otherwise it would store the capitalize version of the given task title.