- Fixed the way it would add tasks
In the addTask(), we have a variable of TaskTitle that has the ternary operator. Now, each time a user inputs a value in the task title input, it would first make the value lowercased and then the capitalizeTitle() function would do its job. This way it would detect duplicates easier because in the localStorage it saves as a capitalize title. And when a user tries to input the same name with different format (i.e., everything in capitals, switches from capitals to lowercases, and others) it would detect it because the given input is a capitalize title.